requirements.txt
scp
wheel
PyYAML
paramiko
pytest-reportlog
labgrid @ git+https://github.com/labgrid-project/labgrid
